split out couple more tests
